# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"NP Computers API","description":"U.S. storefront with simulated payments only. Auth writes require same-origin CSRF token from /api/v1/auth/csrf. Order submission requires Idempotency-Key; declines return PAYMENT_FAILED. Money is USD. Errors use application/problem+json.","version":"1.0.0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/me/profile":{"get":{"tags":["customer-controller"],"operationId":"get","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Profile"}}}}}},"put":{"tags":["customer-controller"],"operationId":"put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Profile"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Profile"}}}}}}},"/api/v1/cart/items/{id}":{"put":{"tags":["order-controller"],"operationId":"quantity","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quantity"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["order-controller"],"operationId":"remove","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/admin/products/{id}":{"put":{"tags":["admin-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/orders":{"get":{"tags":["order-controller"],"operationId":"list","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["order-controller"],"operationId":"place","parameters":[{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaceOrder"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/contact-inquiries":{"post":{"tags":["contact-controller"],"operationId":"post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Inquiry"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/checkout/quote":{"post":{"tags":["order-controller"],"operationId":"quote","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quote"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/reset":{"post":{"tags":["auth-controller"],"operationId":"reset","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reset"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Registration"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","parameters":[{"name":"NP_REFRESH","in":"cookie","required":false,"schema":{"type":"string","default":""}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/recovery":{"post":{"tags":["auth-controller"],"operationId":"recovery","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recovery"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","parameters":[{"name":"NP_REFRESH","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/logout-all":{"post":{"tags":["auth-controller"],"operationId":"all","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credentials"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/admin/users/{id}/roles":{"post":{"tags":["admin-controller"],"operationId":"role","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Role"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/admin/users/{id}/enabled":{"post":{"tags":["admin-controller"],"operationId":"enable","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Enabled"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/admin/products":{"get":{"tags":["admin-controller"],"operationId":"products","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["admin-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/admin/inventory/receipts":{"post":{"tags":["admin-controller"],"operationId":"receipt","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Receipt"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/admin/inventory/adjustments":{"post":{"tags":["admin-controller"],"operationId":"adjust","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Adjustment"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/staff/orders/{id}/status":{"patch":{"tags":["order-controller"],"operationId":"status","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Status"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/products":{"get":{"tags":["catalog-controller"],"operationId":"list_1","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"category","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"manufacturer","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"condition","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"min","in":"query","required":false,"schema":{"type":"number","default":0}},{"name":"max","in":"query","required":false,"schema":{"type":"number","default":9999999999}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/products/{id}":{"get":{"tags":["catalog-controller"],"operationId":"detail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/orders/{id}":{"get":{"tags":["order-controller"],"operationId":"detail_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/cart":{"get":{"tags":["order-controller"],"operationId":"cart","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["order-controller"],"operationId":"clear","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/csrf":{"get":{"tags":["auth-controller"],"operationId":"csrf","parameters":[{"name":"t","in":"query","required":true,"schema":{"$ref":"#/components/schemas/CsrfToken"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/admin/users":{"get":{"tags":["admin-controller"],"operationId":"users","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/admin/mail-sink":{"get":{"tags":["admin-controller"],"operationId":"mail","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/admin/inventory/replenishments":{"get":{"tags":["admin-controller"],"operationId":"replenish","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/admin/audit":{"get":{"tags":["admin-controller"],"operationId":"audit","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"Address":{"type":"object","properties":{"line1":{"type":"string","maxLength":160,"minLength":0},"line2":{"type":"string","maxLength":160,"minLength":0},"city":{"type":"string","maxLength":100,"minLength":0},"state":{"type":"string","minLength":1},"zip":{"type":"string","minLength":1,"pattern":"[0-9]{5}(-[0-9]{4})?"},"country":{"type":"string","minLength":1,"pattern":"US"}},"required":["city","country","line1","state","zip"]},"Contact":{"type":"object","properties":{"firstName":{"type":"string","maxLength":80,"minLength":0},"lastName":{"type":"string","maxLength":80,"minLength":0},"email":{"type":"string","format":"email","maxLength":254,"minLength":0},"phone":{"type":"string","minLength":1,"pattern":"[+0-9 ()-]{7,40}"}},"required":["email","firstName","lastName","phone"]},"Profile":{"type":"object","properties":{"kind":{"type":"string","minLength":1,"pattern":"INDIVIDUAL|BUSINESS"},"businessName":{"type":"string","maxLength":160,"minLength":0},"primary":{"$ref":"#/components/schemas/Contact"},"alternate":{"$ref":"#/components/schemas/Contact"},"mailing":{"$ref":"#/components/schemas/Address"},"shipping":{"$ref":"#/components/schemas/Address"},"version":{"type":"integer","format":"int64"}},"required":["kind","mailing","primary","shipping"]},"Quantity":{"type":"object","properties":{"quantity":{"type":"integer","format":"int32","maximum":1000000,"minimum":1}}},"ProductInput":{"type":"object","properties":{"sku":{"type":"string","maxLength":80,"minLength":0},"manufacturer":{"type":"string","maxLength":80,"minLength":0},"model":{"type":"string","maxLength":120,"minLength":0},"category":{"type":"string","pattern":"PC|LAPTOP|TABLET|ACCESSORY"},"condition":{"type":"string","pattern":"NEW|USED"},"cpu":{"type":"string","maxLength":120,"minLength":0},"ram":{"type":"string","maxLength":80,"minLength":0},"storage":{"type":"string","maxLength":120,"minLength":0},"gpu":{"type":"string","maxLength":120,"minLength":0},"os":{"type":"string","maxLength":120,"minLength":0},"price":{"type":"number","maximum":9999999999.99,"minimum":0.01},"description":{"type":"string","maxLength":2000,"minLength":0},"active":{"type":"boolean"},"version":{"type":"integer","format":"int64"}},"required":["category","condition","description","manufacturer","model","price","sku"]},"PlaceOrder":{"type":"object","properties":{"quoteId":{"type":"string","format":"uuid"},"quoteDigest":{"type":"string","minLength":1},"shippingAddress":{"$ref":"#/components/schemas/Address"},"paymentMethod":{"type":"string","minLength":1,"pattern":"SIMULATED_CARD|SIMULATED_WALLET"},"scenario":{"type":"string","minLength":1,"pattern":"APPROVE|DECLINE"}},"required":["paymentMethod","quoteDigest","quoteId","scenario","shippingAddress"]},"Inquiry":{"type":"object","properties":{"email":{"type":"string","format":"email","maxLength":254,"minLength":0},"phone":{"type":"string","minLength":1,"pattern":"[+0-9 ()-]{7,40}"},"reason":{"type":"string","pattern":"Order Inquiry|Product Inquiry|Support|Cancellation|Other"},"message":{"type":"string","maxLength":500,"minLength":0}},"required":["email","message","phone","reason"]},"Quote":{"type":"object","properties":{"shippingAddress":{"$ref":"#/components/schemas/Address"}},"required":["shippingAddress"]},"Reset":{"type":"object","properties":{"token":{"type":"string","minLength":1},"password":{"type":"string","maxLength":128,"minLength":12}},"required":["password","token"]},"Registration":{"type":"object","properties":{"email":{"type":"string","format":"email","maxLength":254,"minLength":0},"password":{"type":"string","maxLength":128,"minLength":12},"confirmPassword":{"type":"string"}},"required":["confirmPassword","email","password"]},"Recovery":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1}},"required":["email"]},"Credentials":{"type":"object","properties":{"email":{"type":"string","format":"email","maxLength":254,"minLength":0},"password":{"type":"string","maxLength":128,"minLength":12}},"required":["email","password"]},"Role":{"type":"object","properties":{"role":{"type":"string","pattern":"CUSTOMER|ADMIN|SALES|SUPPORT"}},"required":["role"]},"Enabled":{"type":"object","properties":{"enabled":{"type":"boolean"}}},"Receipt":{"type":"object","properties":{"requestId":{"type":"string","format":"uuid"},"reference":{"type":"string","maxLength":80,"minLength":0}},"required":["reference","requestId"]},"Adjustment":{"type":"object","properties":{"productId":{"type":"string","format":"uuid"},"delta":{"type":"integer","format":"int32"},"reason":{"type":"string","maxLength":200,"minLength":0},"reference":{"type":"string","maxLength":80,"minLength":0}},"required":["productId","reason","reference"]},"Status":{"type":"object","properties":{"status":{"type":"string","pattern":"PROCESSING|SHIPPED"},"tracking":{"type":"string","maxLength":160,"minLength":0},"version":{"type":"integer","format":"int64"}},"required":["status"]},"CsrfToken":{"type":"object","properties":{"token":{"type":"string"},"parameterName":{"type":"string"},"headerName":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}